requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.5
django-allauth==65.12.0
django-environ==0.11.2
crispy-bootstrap4==2025.6
Django==4.2.25
django-bootstrap4==24.4
django-crispy-forms==2.4
pillow==10.4.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2